requirements.txt
robotframework
robotframework-appiumlibrary
Appium-Python-Client>=3.0
requests
